refactor(code_gate): collapse __init__ to real facade and drop models shim
- Rewrite `code_gate/__init__.py` so it is the real package contract:
  export the 8 symbols actually used by callers (`run_stage`,
  `StageOptions`, `get_changed_files`, `run_precommit`, `run_check`,
  `run_build`, `detect_language`, `BUCKET_AWARE_STAGES`) and stop
  re-exporting domain value objects (`CheckResult` / `StageOutcome` /
  `Violation`), which already live in `ac_guard.domain.models`.
- Drop `TYPE_EXTENSIONS` from `code_gate.core.__all__` — it is the
  internal data table for `detect_language` and
  `_filter_files_by_type`, with no external consumer.
- Delete `code_gate/models.py` (a backward-compat shim re-exporting
  domain value objects). Production code already imported from
  `ac_guard.domain.models`; only three test files still pointed at
  the shim, all migrated in this PR.
- Migrate `cli/check.py` and `cli/status.py` from
  `from ac_guard.code_gate.core import …` to
  `from ac_guard.code_gate import …` so production callers depend on
  the package facade rather than the implementation submodule.
- Migrate three test files (test_reporter/test_formatting.py,
  test_code_gate_e2e.py, test_m5_report_and_output.py) plus two
  function-local imports inside test_code_gate/test_core.py to
  `ac_guard.domain.models`. Delete test_code_gate/test_models.py —
  its only purpose was validating the shim's re-export path.

Scope is deliberately narrow: this is the first of a planned three-PR
sequence under #109 (reopened to track residuals beyond the reporter
portion already completed in #138). PR2 will migrate
`detect_language` + `TYPE_EXTENSIONS` to a new
`ac_guard.domain.languages` Domain Service. PR3 will relocate the
D3b passthrough handling (commit-msg / pre-merge-commit / pre-rebase
in `cli/check.py:gate_run_command`) into the `cli/` layer and drop
`BUCKET_AWARE_STAGES` from the code_gate facade. The
`gate_run_command` inline `subprocess.run(["pre-commit", "run",
"--hook-stage", …])` block is intentionally untouched here; it moves
in PR3.

No behaviour change: pytest 1041/1041 green, ruff clean, pre-commit
clean (interrogate / import-linter / bandit), import-chain smoke
verified, end-to-end smoke on a fresh project root verified install /
check / gate run --stage pre-commit / gate run --stage commit-msg /
doctor --strict.

Refs #109
Part of #133
